+++ static/app/views/eventsV2/resultsChart.spec.tsx
@@ -4,7 +4,6 @@ import {initializeOrg} from 'sentry-test/initializeOrg';
 import {t} from 'sentry/locale';
 import EventView from 'sentry/utils/discover/eventView';
 import {DisplayModes} from 'sentry/utils/discover/types';
-import {MetricsCardinalityProvider} from 'sentry/utils/performance/contexts/metricsCardinality';
 import ResultsChart from 'sentry/views/eventsV2/resultsChart';
 
 describe('EventsV2 > ResultsChart', function () {
@@ -30,24 +29,33 @@ describe('EventsV2 > ResultsChart', function () {
       projects: [],
     });
     eventView = EventView.fromSavedQueryOrLocation(undefined, location);
+    MockApiClient.addMockResponse({
+      url: '/organizations/org-slug/releases/stats/',
+      body: [],
+    });
+    MockApiClient.addMockResponse({
+      url: '/organizations/org-slug/events-stats/',
+      body: [],
+    });
   });
 
   it('only allows default, daily, previous period, and bar display modes when multiple y axis are selected', function () {
     const wrapper = mountWithTheme(
-      <MetricsCardinalityProvider location={location} organization={organization}>
-        <ResultsChart
-          router={TestStubs.router()}
-          organization={organization}
-          eventView={eventView}
-          location={location}
-          onAxisChange={() => undefined}
-          onDisplayChange={() => undefined}
-          total={1}
-          confirmedQuery
-          yAxis={['count()', 'failure_count()']}
-          onTopEventsChange={() => {}}
-        />
-      </MetricsCardinalityProvider>,
+      <ResultsChart
+        router={TestStubs.router()}
+        disableProcessedBaselineToggle
+        setShowBaseline={() => undefined}
+        showBaseline
+        organization={organization}
+        eventView={eventView}
+        location={location}
+        onAxisChange={() => undefined}
+        onDisplayChange={() => undefined}
+        total={1}
+        confirmedQuery
+        yAxis={['count()', 'failure_count()']}
+        onTopEventsChange={() => {}}
+      />,
       initialData.routerContext
     );
     const displayOptions = wrapper.find('ChartFooter').props().displayOptions;
@@ -67,20 +75,21 @@ describe('EventsV2 > ResultsChart', function () {
 
   it('does not display a chart if no y axis is selected', function () {
     const wrapper = mountWithTheme(
-      <MetricsCardinalityProvider location={location} organization={organization}>
-        <ResultsChart
-          router={TestStubs.router()}
-          organization={organization}
-          eventView={eventView}
-          location={location}
-          onAxisChange={() => undefined}
-          onDisplayChange={() => undefined}
-          total={1}
-          confirmedQuery
-          yAxis={[]}
-          onTopEventsChange={() => {}}
-        />
-      </MetricsCardinalityProvider>,
+      <ResultsChart
+        router={TestStubs.router()}
+        disableProcessedBaselineToggle
+        setShowBaseline={() => undefined}
+        showBaseline
+        organization={organization}
+        eventView={eventView}
+        location={location}
+        onAxisChange={() => undefined}
+        onDisplayChange={() => undefined}
+        total={1}
+        confirmedQuery
+        yAxis={[]}
+        onTopEventsChange={() => {}}
+      />,
       initialData.routerContext
     );
     expect(wrapper.find('NoChartContainer').children().children().html()).toEqual(
@@ -96,20 +105,21 @@ describe('EventsV2 > ResultsChart', function () {
       {field: 'equation|count() + 2'},
     ];
     const wrapper = mountWithTheme(
-      <MetricsCardinalityProvider location={location} organization={organization}>
-        <ResultsChart
-          router={TestStubs.router()}
-          organization={organization}
-          eventView={eventView}
-          location={location}
-          onAxisChange={() => undefined}
-          onDisplayChange={() => undefined}
-          total={1}
-          confirmedQuery
-          yAxis={['count()']}
-          onTopEventsChange={() => {}}
-        />
-      </MetricsCardinalityProvider>,
+      <ResultsChart
+        router={TestStubs.router()}
+        disableProcessedBaselineToggle
+        setShowBaseline={() => undefined}
+        showBaseline
+        organization={organization}
+        eventView={eventView}
+        location={location}
+        onAxisChange={() => undefined}
+        onDisplayChange={() => undefined}
+        total={1}
+        confirmedQuery
+        yAxis={['count()']}
+        onTopEventsChange={() => {}}
+      />,
       initialData.routerContext
     );
     const yAxisOptions = wrapper.find('ChartFooter').props().yAxisOptions;
